roo: fix auth
