+++ src/app/auth.js
@@ -1,0 +1,11 @@
+import { getServerSession } from "next-auth";
+import { handler } from "@/app/api/auth/[...nextauth]/route";
+
+export const getSession = async () => {
+  return await getServerSession(handler);
+};
+
+export const getCurrentUser = async () => {
+  const session = await getSession();
+  return session?.user;
+};
